default to 32 bit integers in sdf index creation functions
