+++ R/sdf_sequence.R
@@ -6,9 +6,10 @@
 #' @param length The desired length of the sequence.
 #' @param repartition The number of partitions to use when distributing the
 #'   data across the Spark cluster.
+#' @param bits The integer size in bits, either 32 or 64.
 #'
 #' @export
-sdf_len <- function(sc, length, repartition = NULL) {
+sdf_len <- function(sc, length, repartition = NULL, bits = c(32, 64)) {
   sdf_seq(sc, 1, length, repartition = repartition)
 }
 
@@ -21,20 +22,30 @@ sdf_len <- function(sc, length, repartition = NULL) {
 #' @param by The increment of the sequence.
 #' @param repartition The number of partitions to use when distributing the
 #'   data across the Spark cluster.
+#' @param bits The integer size in bits, either 32 or 64.
 #'
 #' @export
-sdf_seq <- function(sc, from = 1L, to = 1L, by = 1L, repartition = NULL) {
+sdf_seq <- function(sc, from = 1L, to = 1L, by = 1L, repartition = NULL, bits = c(32, 64)) {
   from <- cast_scalar_integer(from)
   to <- cast_scalar_integer(to + 1)
   by <- cast_scalar_integer(by)
+
+  # validate bits parameter
+  bits <- match.arg(bits)
+
+  type_map <- list(
+    "32" = "Integer",
+    "64" = "Long"
+  )
+  type_name <- type_map[[as.character(type)]]
 
   if (is.null(repartition)) repartition <- invoke(spark_context(sc), "defaultMinPartitions")
   repartition <- cast_scalar_integer(repartition)
 
   rdd <- invoke(spark_context(sc), "range", from, to, by, repartition)
-  rdd <- invoke_static(sc, "sparklyr.WorkerUtils", "mapRddLongToRddRow", rdd)
+  rdd <- invoke_static(sc, "sparklyr.Utils", paste0("mapRdd", type_map[[type]] , "ToRddRow"), rdd)
 
-  schema <- invoke_static(sc, "sparklyr.WorkerUtils", "buildStructTypeForLongField")
+  schema <- invoke_static(sc, "sparklyr.Utils", paste0("buildStructTypeFor", type_map[[type]], "Field"))
   sdf <- invoke(hive_context(sc), "createDataFrame", rdd, schema)
 
   sdf_register(sdf)
@@ -48,8 +59,9 @@ sdf_seq <- function(sc, from = 1L, to = 1L, by = 1L, repartition = NULL) {
 #' @param along Takes the length from the length of this argument.
 #' @param repartition The number of partitions to use when distributing the
 #'   data across the Spark cluster.
+#' @param bits The integer size in bits, either 32 or 64.
 #'
 #' @export
-sdf_along <- function(sc, along, repartition = NULL) {
-  sdf_len(sc, length(along), repartition = repartition)
+sdf_along <- function(sc, along, repartition = NULL, bits = c(32, 64)) {
+  sdf_len(sc, length(along), repartition = repartition, bits = bits)
 }
