+++ inst/livy/spark-2.3.0/livyutils.scala
@@ -13,7 +13,6 @@ object LivyUtils {
     "ArrowConverters" -> ArrowConverters,
     "ApplyUtils" -> ApplyUtils,
     "WorkerHelper" -> WorkerHelper,
-    "WorkerUtils" -> WorkerUtils,
     "MLUtils" -> MLUtils,
     "MLUtils2" -> MLUtils2,
     "BucketizerUtils" -> BucketizerUtils
